fix(frontend)[OK-13] Rename other file names and function names
