+++ frontend/src/components/board/SidebarItem.tsx
